requirements/base.txt
asgiref==3.2.3
Django==2.2.9
django-appconf==1.0.3
django-compressor==2.4
django-summernote==0.8.11.6
django-taggit==1.2.0
libsass==0.19.4
Pillow==7.0.0
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
